fully implemented math lib. collision back to working order. triangles tracked properly (multiple cubes can be in the scene at once)

+++ project-assist-build-version-0.5-indev/Light.h
@@ -12,28 +12,28 @@ public:
 	struct LightData {
 		
 		struct {
-			double x{ 0.0f };
-			double y{ 0.0f };
-			double z{ 0.0f };
-			double w{ 1.0f };
+			float x{ 0.0f };
+			float y{ 0.0f };
+			float z{ 0.0f };
+			float w{ 1.0f };
 		} pos;
 		struct {
-			double x{ 1.0f };
-			double y{ 1.0f };
-			double z{ 1.0f };
-			double w{ 1.0f };
+			float x{ 1.0f };
+			float y{ 1.0f };
+			float z{ 1.0f };
+			float w{ 1.0f };
 		} direction;
 		struct {
-			double r{ 1.0f };
-			double g{ 1.0f };
-			double b{ 1.0f };
-			double a{ 1.0f };
+			float r{ 1.0f };
+			float g{ 1.0f };
+			float b{ 1.0f };
+			float a{ 1.0f };
 		} color;
 
-		double spotAngle{ 15.0f };
-		double constAtten{ 1.0f };
-		double linAtten{ 0.08f };
-		double quadAtten{ 0.0f };
+		float spotAngle{ 15.0f };
+		float constAtten{ 1.0f };
+		float linAtten{ 0.08f };
+		float quadAtten{ 0.0f };
 
 		int type{ POINT_LIGHT };
 		int isEnabled{ false };
