+++ project-assist-build-version-0.5-indev/Material.h
@@ -66,12 +66,12 @@ public:
 	bool isMaterialKAXYZ{ false };
 
 	// Material KA values
-	std::vector<double> KA_RGB{0.0f, 0.0f, 0.0f};
+	std::vector<float> KA_RGB{0.0f, 0.0f, 0.0f};
 
 	std::string KA_file{};
 	double KA_factor{};
 
-	std::vector<double> KA_XYZ{ 0.0f, 0.0f, 0.0f };
+	std::vector<float> KA_XYZ{ 0.0f, 0.0f, 0.0f };
 
 	// Material KD type
 	bool isMaterialKDRGB{ false };
@@ -82,12 +82,12 @@ public:
 	bool isMaterialKDXYZ{ false };
 
 	// Material KD values
-	std::vector<double> KD_RGB{ 0.0f, 0.0f, 0.0f };
+	std::vector<float> KD_RGB{ 0.0f, 0.0f, 0.0f };
 
 	std::string KD_file{};
 	double KD_factor{};
 
-	std::vector<double> KD_XYZ{ 0.0f, 0.0f, 0.0f };
+	std::vector<float> KD_XYZ{ 0.0f, 0.0f, 0.0f };
 
 	// Material KS type
 	bool isMaterialKSRGB{ false };
@@ -98,10 +98,10 @@ public:
 	bool isMaterialKSXYZ{ false };
 
 	// Material KS values
-	std::vector<double> KS_RGB{ 0.0f, 0.0f, 0.0f };
+	std::vector<float> KS_RGB{ 0.0f, 0.0f, 0.0f };
 
 	std::string KS_file{};
-	double KS_factor{};
+	float KS_factor{};
 
 	std::vector<double> KS_XYZ{ 0.0f, 0.0f, 0.0f };
 
@@ -114,28 +114,28 @@ public:
 	bool isMaterialTFXYZ{ false };
 
 	// Material TF values
-	std::vector<double> TF_RGB{ 0.0f, 0.0f, 0.0f };
+	std::vector<float> TF_RGB{ 0.0f, 0.0f, 0.0f };
 
 	std::string TF_file{};
-	double TF_factor{};
+	float TF_factor{};
 
-	std::vector<double> TF_XYZ{ 0.0f, 0.0f, 0.0f };
+	std::vector<float> TF_XYZ{ 0.0f, 0.0f, 0.0f };
 
 	// Illumination Flags
 	std::vector<bool> isIllumFlags{ false, false, false, false, false, false, false, false, false, false, false };
 
 	// D
-	double D_factor{};
+	float D_factor{};
 	bool isDHalo{ false };
 
 	// Material NS values
-	double NS_exponent{};
+	float NS_exponent{};
 
 	// Material Sharpness
-	double sharpness{};
+	float sharpness{};
 
 	// Material NI values
-	double NI_optical_density{};
+	float NI_optical_density{};
 
 	// Material Textures /* Implement later */
